# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Rokwire Content Building Block API Documentation.",
        "title": "Rokwire Content Building Block API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.1.10"
    },
    "host": "localhost",
    "basePath": "/content",
    "paths": {
...
        "/content_item/categories": {
            "get": {
                "security": [
                    {
                        "UserAuth": []
                    }
                ],
                "description": "Retrieves  all content item categories that have in the database",
                "tags": [
                    "Client"
                ],
                "operationId": "GetContentItemsCategories",
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/content_items": {
            "get": {
                "security": [
                    {
                        "UserAuth": []
                    }
                ],
                "description": "Retrieves  all content items. \u003cb\u003e The data element could be either a primitive or nested json or array.\u003c/b\u003e",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "operationId": "GetContentItems",
                "parameters": [
                    {
                        "type": "string",
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "limit - limit the result",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "order - Possible values: asc, desc. Default: desc",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "description": "Optional - body json of the all items ids that need to be filtered. NOTE: Bad/broken json will be interpreted as an empty filter and the request will be proceeded further.",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/getContentItemsRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ContentItem"
                            }
                        }
                    }
                }
            }
        },
        "/content_items/{id}": {
            "get": {
                "security": [
                    {
                        "UserAuth": []
                    }
                ],
                "description": "Retrieves a content item by id. \u003cb\u003e The data element could be either a primitive or nested json or array.\u003c/b\u003e",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "operationId": "GetContentItem",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ContentItem"
                        }
                    }
                }
            }
        },
...
        "/image": {
            "post": {
                "security": [
                    {
                        "UserAuth": []
                    }
                ],
                "description": "Uploads an image to AWS S3",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "operationId": "UploadImage",
                "parameters": [
                    {
                        "description": "path - path within the S3 bucket",
                        "name": "path",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "width - width of the image to resize. If width and height are missing - then the new image will use the original size",
                        "name": "width",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "height - height of the image to resize. If width and height are missing - then the new image will use the original size",
                        "name": "height",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "quality - quality of the image. Default: 100",
                        "name": "quality",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "fileName - the uploaded file name",
                        "name": "fileName",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/profile_photo": {
            "get": {
                "security": [
                    {
                        "RokwireAuth": []
                    }
                ],
                "description": "Deletes the profile photo of the user who request",
                "tags": [
                    "Client"
                ],
                "operationId": "DeleteProfilePhoto",
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "RokwireAuth": []
                    }
                ],
                "description": "Stores profile photo",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "operationId": "StoreProfilePhoto",
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/profile_photo/{user-id}": {
            "get": {
                "security": [
                    {
                        "RokwireAuth": []
                    }
                ],
                "description": "Retrieves the profile photo",
                "tags": [
                    "Client"
                ],
                "operationId": "GetProfilePhoto",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Possible values: default, medium, small",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/student_guides": {
            "get": {
                "security": [
                    {
                        "RokwireAuth": []
                    }
                ],
                "description": "Retrieves  all student guides",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "operationId": "GetStudentGuides",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Coma separated IDs of the desired records",
                        "name": "ids",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/student_guides/{id}": {
            "get": {
                "security": [
                    {
                        "RokwireAuth": []
                    }
                ],
                "description": "Retrieves a student guide by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "operationId": "GetStudentGuide",
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/twitter/users/{user_id}/tweets": {
            "get": {
                "security": [
                    {
                        "RokwireAuth": []
                    }
                ],
                "description": "Retrieves Twitter tweets for the specified user id. This API is intended to be invoked with the original Twitter query params to https://api.twitter.com/2/users/%s/tweets",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "operationId": "GetTweeterPosts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ContentItem": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "data": {
                    "description": "could be eigther a primitive or nested json or array"
                },
                "date_created": {
                    "type": "string"
                },
                "date_updated": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "getContentItemsRequestBody": {
            "type": "object",
            "properties": {
                "categories": {
                    "description": "List of Categories for the filter. Optional and may be null or missing.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ids": {
                    "description": "List of IDs for the filter. Optional and may be null or missing.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminGroupAuth": {
            "type": "apiKey",
            "name": "GROUP",
            "in": "header"
        },
        "AdminUserAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header (add Bearer prefix to the Authorization value)"
        },
        "RokwireAuth": {
            "type": "apiKey",
            "name": "ROKWIRE-API-KEY",
            "in": "header"
        },
        "UserAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header (add Bearer prefix to the Authorization value)"
        }
    }
}